requirements.txt
python-fasthtml
uvicorn>=0.29
google-generativeai